requirements.txt
google-generativeai>=0.8.0
python-dotenv>=1.0.0
chromadb>=0.4.0
openai>=1.0.0
anthropic>=0.7.0
